\name{gpuExpCpp}
\alias{gpuExpCpp}


\title{
Just a test.
}
\description{
Just a test.
}
\usage{
gpuExpCpp(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
An integer vector.
}
  \item{y}{
An integer vector.
}
}
\details{
Just a test.
}
\value{
Just a test.
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


\examples{
gpuExpCpp(1L : 1024L, 1L : 1024L)
}
